Moved class to new file
